+++ tests/integration/clean_db.py
@@ -8,10 +8,8 @@ import asyncio
 import logging
 import os
 
-from motor.motor_asyncio import AsyncIOMotorClient
+from mongo import Mongo
 
-serverTimeout = 15000
-connectTimeout = 15000
 
 # === Global vars ===
 DATABASE = os.getenv("MONGO_DATABASE", "default")
@@ -23,48 +21,23 @@ LOG = logging.getLogger(__name__)
 LOG.setLevel(logging.DEBUG)
 
 
-def create_db_client(url: str) -> AsyncIOMotorClient:
-    """Initialize database client for AioHTTP App.
-
-    :returns: Coroutine-based Motor client for Mongo operations
-    """
-    return AsyncIOMotorClient(url, connectTimeoutMS=connectTimeout, serverSelectionTimeoutMS=serverTimeout)
-
-
-async def purge_mongodb(url: str) -> None:
-    """Erase database."""
-    client = create_db_client(url)
-    LOG.debug(f"current databases: {*await client.list_database_names(),}")
-    LOG.debug("=== Drop curent database ===")
-    await client.drop_database(DATABASE)
-    LOG.debug("=== DONE ===")
-
-
-async def clean_mongodb(url: str) -> None:
-    """Clean Collection and recreate it."""
-    client = create_db_client(url)
-    db = client[DATABASE]
-    LOG.debug(f"Database to clear: {DATABASE}")
-    collections = await db.list_collection_names()
-    LOG.debug(f"=== Collections to be cleared: {collections} ===")
-    LOG.debug("=== Delete all documents in all collections ===")
-    for col in collections:
-        x = await db[col].delete_many({})
-        LOG.debug(f"{x.deleted_count}{' documents deleted'}\t{'from '}{col}")
-    LOG.debug("=== DONE ===")
-
-
 if __name__ == "__main__":
-    parser = argparse.ArgumentParser(description="Process some integers.")
+    parser = argparse.ArgumentParser(description="Clear or recreate mongo db.")
     parser.add_argument("--tls", action="store_true", help="add tls configuration")
     parser.add_argument("--purge", action="store_true", help="destroy database")
+    parser.add_argument("--recreate", action="store_true", help="Recreate database")
     args = parser.parse_args()
+
     url = f"mongodb://{AUTHDB}:{AUTHDB}@{HOST}/{DATABASE}?authSource=admin"
     if args.tls:
         _params = "?tls=true&tlsCAFile=./config/cacert&tlsCertificateKeyFile=./config/combined"
         url = f"mongodb://{AUTHDB}:{AUTHDB}@{HOST}/{DATABASE}{_params}&authSource=admin"
     LOG.debug(f"=== Database url {url} ===")
+
+    mongo = Mongo(url)
     if args.purge:
-        asyncio.run(purge_mongodb(url))
+        asyncio.run(mongo.drop_db())
+    elif args.recreate:
+        asyncio.run(mongo.recreate_db())
     else:
-        asyncio.run(clean_mongodb(url))
+        asyncio.run(mongo.clean_db())